Add files via upload
Supermarket Sales Analysis

This project involves the analysis of a supermarket's sales data, with a focus on understanding key aspects of its operations, customer behavior, and product performance. The dataset used in this project includes several crucial columns that capture various dimensions of the supermarket's business.

Row ID: A unique identifier for each row in the dataset.
Order ID: A unique identifier for each order placed in the supermarket.
Order Date: The date on which the order was placed.
Ship Date: The date on which the order was shipped to the customer.
Ship Mode: The mode of shipping used for the order (e.g., standard, expedited).
Customer ID: A unique identifier for each customer.
Customer Name: The name of the customer who placed the order.
Segment: The customer segment (e.g., Consumer, Corporate, Home Office) to which the customer belongs.
Country: The country in which the customer resides.
City: The city where the customer is located.
State: The state in which the customer is located.
Postal Code: The postal code for the customer's location.
Region: The region within the country where the customer is located.
Product ID: A unique identifier for each product.
Category: The high-level category to which the product belongs (e.g., Technology, Office Supplies, Furniture).
Sub-Category: A more specific category within the broader category (e.g., Phones, Chairs).
Product Name: The name of the product purchased.
Sales: The total sales value of the product.
Quantity: The number of units of the product sold.
Discount: The discount applied to the product.
Profit: The profit made from the sale of the product.